requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
requests==2.28.2
pytube==12.1.3
numpy==1.24.2
opencv-python==4.7.0.72
mysql==0.0.3 
mysqlclient==2.1.1
mysql-connector==2.2.9
